requirements.txt
mss>=9.0.1
Pillow>=10.0.0
numpy>=1.24.0
pywebview>=4.4.1
pystray>=0.19.5
pywin32>=306
